+++ orchestrator/nodelog_mask.go
@@ -1,7 +1,7 @@
 package orchestrator
 
 import (
-	"github.com/sparkwing-dev/sparkwing/secrets"
+	"github.com/sparkwing-dev/sparkwing/internal/secrets"
 	"github.com/sparkwing-dev/sparkwing/sparkwing"
 )